Move difficulty related functions to its own package
